requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
healthcheck==1.3.3
GeoAlchemy2==0.13.3
graypy==2.1.0
pytz==2023.3